// src/generated/model/batchResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { BatchResponseStatus } from './batchResponseStatus.ts';

export interface BatchResponse {
  id?: string;
  intakeId?: string;
  code?: string;
  name?: string;
  startDate?: string;
  endDate?: string;
  status?: BatchResponseStatus;
  remarks?: string;
  createdAt?: string;
  updatedAt?: string;
}